+++ SciDataViewer/shark/Skeleton.cpp
@@ -28,10 +28,12 @@ void Bone::setApplyParentTransform(bool toggle)
 {
 	applyParentTransform = toggle;
 }
+
 Skeleton::Skeleton()
 {
 
 }
+
 Skeleton::~Skeleton()
 {
 
@@ -39,14 +41,15 @@ Skeleton::~Skeleton()
 
 int Skeleton::getNumBones()
 {
-	return bones.size();
+	return (int) bones.size();
 }
+
 void Skeleton::importBonesFromAssimp(aiNode * node, BoneTreeNode & thisNode)
 {
 	int boneIdx = addBone(std::string(node->mName.data), glm::transpose(glm::make_mat4(&node->mTransformation.a1)));
 
 	thisNode.boneIdx = boneIdx;
-	for (int i = 0; i < node->mNumChildren; i++)
+	for (unsigned int i = 0; i < node->mNumChildren; i++)
 	{
 		thisNode.children.push_back(BoneTreeNode());
 		importBonesFromAssimp(node->mChildren[i], thisNode.children.back());
@@ -58,7 +61,7 @@ void Skeleton::importBonesFromAssimp(aiNode * node)
 {
 	rootInverseTransform = glm::transpose(glm::make_mat4(&node->mTransformation.a1));
 	rootInverseTransform = glm::inverse(rootInverseTransform);
-	for (int i = 0; i < node->mNumChildren; i++)
+	for (unsigned int i = 0; i < node->mNumChildren; i++)
 	{
 		boneRoot.children.push_back(BoneTreeNode());
 		importBonesFromAssimp(node->mChildren[i], boneRoot.children.back());
@@ -84,7 +87,7 @@ Bone * const Skeleton::getBone(std::string boneName)
  */
 int Skeleton::addBone(std::string boneName, glm::mat4 boneMtx)
 {
-	Bone bone(boneName, bones.size(), boneMtx);
+	Bone bone(boneName, (int) bones.size(), boneMtx);
 	bones.push_back(bone);
 	boneMap[boneName] = bone.getIndex();
